+++ frontend/app/src/typing/types.ts
@@ -21,11 +21,7 @@ export interface GeneralSettings {
   readonly ksmRpcEndpoint: string;
   readonly balanceSaveFrequency: number;
   readonly dateDisplayFormat: string;
-  readonly thousandSeparator: string;
-  readonly decimalSeparator: string;
-  readonly currencyLocation: CurrencyLocation;
   readonly selectedCurrency: Currency;
-  readonly krakenAccountType: string;
   readonly activeModules: SupportedModules[];
   readonly btcDerivationGapLimit: number;
   readonly displayDateInLocaltime: boolean;
@@ -58,8 +54,9 @@ export interface Credentials {
 export type ExchangeRates = { [key: string]: BigNumber };
 
 export interface ExchangeInfo {
-  readonly name: string;
+  readonly location: string;
   readonly balances: AssetBalances;
+  readonly total: BigNumber;
 }
 
 export type ExchangeData = { [exchange: string]: AssetBalances };